+++ src/features/UserPosts/index.type.tsx
@@ -1,0 +1,10 @@
+export interface Post {
+	id: number;
+	name: string;
+	email: string;
+	createdAt: string;
+}
+
+export interface responseData {
+	error?: string;
+}
